changed return value of GetNumScoreComponents() to size_t to match virt.func.declaration in ScoreProducer.h
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@896 1f5c12ca-751b-0410-a591-d2e778427230

+++ trunk/moses/src/DummyScoreProducers.h
@@ -15,7 +15,7 @@ public:
 
 	float CalculateDistortionScore(const WordsRange &prev, const WordsRange &curr) const;
 
-	unsigned int GetNumScoreComponents() const;
+	size_t GetNumScoreComponents() const;
 	const std::string GetScoreProducerDescription() const;
 };
 
@@ -26,7 +26,7 @@ class WordPenaltyProducer : public ScoreProducer {
 public:
 	WordPenaltyProducer();
 
-	unsigned int GetNumScoreComponents() const;
+	size_t GetNumScoreComponents() const;
 	const std::string GetScoreProducerDescription() const;
 };
 
@@ -35,7 +35,7 @@ class UnknownWordPenaltyProducer : public ScoreProducer {
 public:
 	UnknownWordPenaltyProducer();
 
-	unsigned int GetNumScoreComponents() const;
+	size_t GetNumScoreComponents() const;
 	const std::string GetScoreProducerDescription() const;
 };
